IMR-12 refactor : apart daliy chart and song info from song detail crawler
